Refactor ui elements to new package. Including test navigation between home and details fragments.
