+++ app/src/main/res/layout/fragment_details.xml
@@ -3,7 +3,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".details.DetailsFragment">
+    tools:context=".ui.details.DetailsFragment">
+
+    <Button
+        android:id="@+id/back"
+        android:text="Go to home"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end"
+        />
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
