+++ app/src/main/res/layout/fragment_home.xml
@@ -3,7 +3,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.HomeFragment">
+    tools:context=".ui.home.HomeFragment">
+
+    <Button
+        android:id="@+id/next"
+        android:text="Go to details"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end"
+        />
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
