+++ app/src/main/res/navigation/nav_graph.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/Home"
-        android:name="com.example.townsqktacademy.home.HomeFragment"
+        android:name="com.example.townsqktacademy.ui.home.HomeFragment"
         android:label="Home"
         tools:layout="@layout/fragment_home">
 
@@ -18,7 +18,7 @@
     </fragment>
     <fragment
         android:id="@+id/Details"
-        android:name="com.example.townsqktacademy.details.DetailsFragment"
+        android:name="com.example.townsqktacademy.ui.details.DetailsFragment"
         android:label="Details"
         tools:layout="@layout/fragment_details">
